requirements.txt
backcall==0.1.0
decorator==4.3.0
Django==2.0.6
djangorestframework==3.8.2
Faker==0.8.16
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2-binary==2.7.5
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.4
simplegeneric==0.8.1
six==1.11.0
text-unidecode==1.2
traitlets==4.3.2
wcwidth==0.1.7
